refactor: Rename internal package to pkg
Fixes #57
