+++ pkg/ansi/init_windows.go
