+++ pkg/output/aws_credentials_file_test.go
@@ -21,7 +21,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/okta/okta-aws-cli/internal/aws"
+	"github.com/okta/okta-aws-cli/pkg/aws"
 	"github.com/stretchr/testify/assert"
 )
