+++ pkg/sessiontoken/multiple_fed_apps_error.go
