+++ pkg/sessiontoken/sessiontoken_test.go
